requirements.txt
asgiref==3.8.1
boto3==1.35.26
botocore==1.35.26
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crypto==1.4.1
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.14
django-cors-headers==4.4.0
django-ses==4.1.1
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
jmespath==1.0.1
jsonpickle==3.3.0
mssql-django==1.5
Naked==0.1.32
oauthlib==3.2.2
pesepay==1.0.8
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
sendgrid==6.11.0
shellescape==3.8.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
